ai_engine/3dgs/src/libs/Depth-Anything-3/pyproject.toml
[build-system]
requires = ["hatchling>=1.25", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "depth-anything-3"
version = "0.0.0"
description = "Depth Anything 3"
readme = "README.md"
requires-python = ">=3.9, <=3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Your Name" }]

dependencies = [
    "pre-commit",
    "trimesh",
    "torch>=2",
    "torchvision",
    "einops",
    "huggingface_hub",
    "imageio",
    "numpy<2",
    "opencv-python",
    "xformers",
    "open3d",
    "fastapi",
    "uvicorn",
    "requests",
    "typer",
    "pillow",
    "omegaconf",
    "evo",
    "e3nn",
    "moviepy",
    "plyfile",
    "pillow_heif",
    "safetensors",
    "uvicorn",
    "moviepy",
    "typer>=0.9.0",
    "pycolmap",
]

[project.optional-dependencies]
app = ["gradio>=5", "pillow>=9.0"] # requires that python3>=3.10
gs = ["gsplat @ git+https://github.com/nerfstudio-project/gsplat.git@0b4dddf04cb687367602c01196913cde6a743d70"]
all = ["depth-anything-3[app,gs]"]


[project.scripts]
da3 = "depth_anything_3.cli:app"

[project.urls]
Homepage = "https://github.com/ByteDance-Seed/Depth-Anything-3"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/depth_anything_3"]

[tool.hatch.build.targets.sdist]
include = [
  "/README.md",
  "/pyproject.toml",
  "/src/depth_anything_3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
plugins = ["jaxtyping.mypy_plugin"]

[tool.black]
line-length = 99
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["bson","cruise","cv2","dataloader","diffusers","omegaconf","tensorflow","torch","torchvision","transformers","gsplat"]
known_first_party = ["common", "data", "models", "projects"]
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
skip_gitignore = true
line_length = 99
no_lines_before="THIRDPARTY"
